server/requirements.txt
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
psycopg2-binary>=2.9.0,<3.0.0
paho-mqtt>=1.6.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
